pyproject.toml
[project]
name = "real-time-sales-forcast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas",
    "numpy",
    "faiss-cpu",
    "transformers",
    "torch",
    "sentence-transformers>=5.4.1",
    "fastapi>=0.136.0",
    "uvicorn>=0.46.0",
    "streamlit>=1.56.0",
    "scikit-learn>=1.8.0",
    "tqdm>=4.67.3",
    "mcp>=1.27.0",
    "geopy>=2.4.1",
]
